Improved the performance of "GetNextValue()" subfunction withinb "DrawText()" which is now heap alloc free, which saves TONS of allocs by being called 60x per sec

+++ QuestHandler.cs
@@ -126,7 +126,7 @@ public struct AllStats : IComparable<AllStats>
     }
 }
 
-public readonly record struct SubQuest(int Count, double Interval)  
+public readonly record struct SubQuest(int Count, float Interval)  
 {
     public int CompareTo(EventStats other)
     {
